New functionality: POB restricted commodity list
New functionality - forbidden commodity list on POBs. Blocks attempts to sell a given cargo on a POB regardless if it's already on the manifest, base administrator status and so on. Allows to purchase the commodity from the base if it had it onboard due to admin intervention or had it stashed before this change went live.

new .cfg file: base_forbidden_cargo.cfg
structure:
[forbidden_commodities]

commodity_name = commodity_passengers
commodity_name = commodity_cardamine